big polish sweep
- tweaked the parameters of many game objects and settings
- renamed a few things
- skip the win screen with enter/space/click
- added shift+F to enter fullscreen
- made data types for menu mode + difficulty selection.
  - currently 1 player is default until the menu is implemented. just gotta recompile to change until then lol
- removed compiler warnings by properly casting
- use floats for internal raylib logo animation
  - also made the draw animation function cleaner

+++ code/config.h
@@ -1,3 +1,6 @@
+// EXPLANATION:
+// This file is for configuring aspects of the program outside of game logic
+
 #ifndef TEST_CONFIG_HEADER_GUARD
 #define TEST_CONFIG_HEADER_GUARD
 
@@ -8,13 +11,12 @@
 #define ASPECT_RATIO (4.0f / 3.0f)
 #define RENDER_HEIGHT 1080 // The native resolution of the game itself.
                            // This will scale to the game window.
-#define RENDER_WIDTH (RENDER_HEIGHT * ASPECT_RATIO)
+#define RENDER_WIDTH (int)(RENDER_HEIGHT * ASPECT_RATIO)
 
 #define DEFAULT_HEIGHT 720 // Default size of the game window.
-#define DEFAULT_WIDTH (DEFAULT_HEIGHT * ASPECT_RATIO)
+#define DEFAULT_WIDTH (int)(DEFAULT_HEIGHT * ASPECT_RATIO)
 
-// NOTE: Very high fps (~500+) is buggy because delta time between frames is too low to be tracked precisely.
-// Unsure if/how to fix this so that the game could theoretically run fine at obsurdly high framerates.
+// there may be small bugs with very high FPS (uncapped + no vsync), but should work fine overall
 #define MAX_FRAMERATE 120 // Set to 0 for uncapped framerate
 #define VSYNC_ENABLED true
