+++ code/logo.h
@@ -5,7 +5,7 @@
 
 // Macros
 #define RAYLIB_LOGO_WIDTH 512 // takes multiples of 64
-                              // 192 is the minimum size
+                                 // 192 is the minimum size
 #define RAYLIB_LOGO_OUTLINE (RAYLIB_LOGO_WIDTH / 16)
 #define RAYLIB_LOGO_FONT_SIZE (RAYLIB_LOGO_WIDTH/8 + RAYLIB_LOGO_OUTLINE)
 
@@ -13,17 +13,17 @@
 typedef enum LogoState { START, GROW1, GROW2, TEXT, PAUSE, END } LogoState;
 
 typedef struct Logo {
-    int positionX;
-    int positionY;
+    float positionX;
+    float positionY;
 
     float elapsedTime;
     int lettersCount;
 
-    int topSideRecWidth;
-    int leftSideRecHeight;
+    float topSideRecWidth;
+    float leftSideRecHeight;
 
-    int bottomSideRecWidth;
-    int rightSideRecHeight;
+    float bottomSideRecWidth;
+    float rightSideRecHeight;
 
     LogoState state;   // Tracking animation states (State Machine)
     float alpha; // Useful for fading
